+++ app/build.gradle
@@ -61,4 +61,5 @@ dependencies {
     androidTestImplementation 'androidx.compose.ui:ui-test-junit4:1.3.3'
     debugImplementation 'androidx.compose.ui:ui-tooling:1.3.3'
     debugImplementation 'androidx.compose.ui:ui-test-manifest:1.3.3'
+    implementation 'com.google.devtools.ksp:symbol-processing-api:1.8.0-1.0.8'
 }
