+++ app/src/main/java/com/example/myapplication/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.theme
+package com.example.myapplication.presentation.theme
 
 import androidx.compose.ui.graphics.Color
